Enabled working services and moved URL share button
